+++ packages/@cdktf/provider-generator/lib/get/generator/emitter/attributes-emitter.ts
@@ -1,6 +1,6 @@
 import { CodeMaker } from "codemaker";
 import { AttributeModel } from "../models";
-import { downcaseFirst } from "../../../util";
+import { downcaseFirst, uppercaseFirst } from "../../../util";
 import { CUSTOM_DEFAULTS } from "../custom-defaults";
 
 function titleCase(value: string) {
@@ -143,6 +143,11 @@ export class AttributesEmitter {
     }
     if (type.isBoolean) {
       return `this.getBooleanAttribute('${att.terraformName}') as any`;
+    }
+    if (type.isMap) {
+      return `this.get${uppercaseFirst(att.mapType)}MapAttribute('${
+        att.terraformName
+      }')`;
     }
     if (process.env.DEBUG) {
       console.error(